ParametricSurface replaces the role of SurfaceMobject, and instead of tracking normals infinitesimal nudges are tracked
